testing git hook

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/01 12:33:18 by vviterbo          #+#    #+#             */
-/*   Updated: 2024/11/06 17:30:24 by vviterbo         ###   ########.fr       */
+/*   Updated: 2024/11/06 17:44:15 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,7 @@
 # include <stdlib.h>
 # include <unistd.h>
 # include <float.h>
+# include <stdbool.h>
 
 enum Datatype {
 	SHORT,
@@ -93,8 +94,8 @@ int		ft_isnumber(char *str);
 double	ft_min(double a, double b);
 double	ft_max(double a, double b);
 int		ft_lst_isin(t_list **lst, void *elem, int size);
-float	ft_lstmin(t_list **lst, int dtype);
-float	ft_lstmax(t_list **lst, int dtype);
+double	ft_lstmin(t_list **lst, int dtype);
+double	ft_lstmax(t_list **lst, int dtype);
 char	**ft_strarray_mapi(char **strarray, char *(*f)(const char *));
 size_t	ft_arrlen(char **array);
 void	ft_free_array(void **array, int size);
